requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
PyMySQL==0.9.3
